2lab/requirements.txt
fastapi==0.111.0
uvicorn==0.21.1
sqlalchemy==2.0.36
python-jose[cryptography]==3.3.0
bcrypt==4.1.3
python-dotenv==1.0.0
pydantic-settings==2.0.3